Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-2
> 
> Still fails to deploy
> https://logs.onap.org/onap-integration/gating//128222-4/k8s/k8s/kubernetes-status/pod-onap-vfc-vnfmgr-7995756757-s885t-vfc-vnfmgr-logs.html

I checked all the gating results at:https://logs.onap.org/onap-integration/gating/?C=M;O=A.
vfc-vnfmgr works in the most of th results. It seems the env is not stable. In this version, only mirror change made for this project and it will not affect the healthcheck or any test. I can't figure out any reason for this failure. Could you give some tips for why it works in most of the gating results, but fails in the some gating results.

Patch-set: 5