Update patch set 10
Patch Set 10:

(1 comment)

In particular reference to the GUI components:

The config map should be altered slightly
data:
{{ tpl (.Files.Glob "resources/config/*.{xml,yaml,yml}").AsConfig . | indent 2 }}

An environment substitution init container needs to be added along with a config file to allow gui to be configured for http(s) as well as change the other config parameters.
The resultant values should then be mounted to the main container.
---
Cert Initializer references should be removed from the deployment if service mesh is active.

Please feel free to contact if you have further questions

Patch-set: 10