Update patch set 2
Patch Set 2: Code-Review-1

Can we discuss this in our weekly meeting Sirisha?   Not sure how this is going to work if some of our interacting components are moved to strimzi and others aren't (as is in this review).

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1