Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Just curious to know why this patch was abandoned in favor of new https://gerrit.onap.org/r/c/oom/+/130171?

Oh it's kind of a long story. I met a problem when git commit, there were 2 change id, and warned that i was 2 commit before master. So i deleted the local repo, and repull patch3. But this time it said another local proj is also modifing on this proj (I think it mean the one i deleted). So I abandoned this one.

Patch-set: 3