Update patch set 19
Patch Set 19:

Yes, I see these NodePort conflicts appearing from time to time.
The problem seems, that there is a bug in the strimzi chart.
https://git.onap.org/oom/tree/kubernetes/strimzi/templates/strimzi-kafka.yaml#n49
As you can see, it reserves reserves the ports:
30491: onap-strimzi-kafka-external-bootstrap
30492: onap-strimzi-kafka-0
30493: onap-strimzi-kafka-1 
The port for onap-strimzi-kafka-2 is missing and therefor randomly chosen !!
If this is in conflict with any other service, like here a1policymanagement, this fails.
So the simple solution is to add the lines:
            - broker: 2
              nodePort: {{ .Values.global.nodePortPrefixExt }}94

but check, if the port is not already reserved.
I will create a ticket for Strimzi... and provide a fix.

Patch-set: 19
Reviewer: Gerrit User 2679 <2679@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>