Update patch set 36
Patch Set 36:

> Patch Set 36:
> 
> > Patch Set 36:
> > 
> > > Patch Set 36:
> > > 
> > > @liam - if you would agree, I disabled bringing up policy-gui component so to make progress on this.   When I installed this patch, I see more than one issue (for example filebeat sidecar is not disabled) that I think is independent of this review, and would need to be looked at more in detail.   I didn't see the issue with the pap that shows here in the last report after a few deploys/undeploys.  Let's try again.
> > 
> > @Jorge, I'll ask Fiachra to see if he's happy to let this through now. We can deal with policy/gui in another commit if we decide to re-enable it.
> 
> @Jorge @Liam If the failure of pap is indeed a password complexity issue - it relates to the passwords generated by the common policy chart. I would guess that sometimes the passwords generated will be fine (because they don't contain troublesome characters) and sometimes they will not. Each time you redeploy a new password will be generated (correct me if I am wrong). This issue is known in oom. It may cause intermittent failures here.
> 
> CPS and other components use the values.yaml parameter "passwordStrengthOverride = basic" to avoid these troublesome characters in generated passwords. If indeed this error is caused by passwords characters, this should permanently fix the issue.

@saul I believe there are several other issues in the GUI unrelated to password.   
The PAP issue (which does not happen now) was the one that relates to the password and that seemed to have been sporadic.   When I played it locally, I don't see it, and don't see it in the latest run (or most of the other gating runs for this review).
With the GUI when I brought it up yesterday, it was crashing with other errors related to certificate staff, not being able to find the keystore if I recall correctly (it was null).    There are other things, that just doesn't seem right, as having filebeat running by default trying to send logs to a non-existing server.    Provided that GUI doesn't seem used, better to deal with the issues of that component apart from this review IMO.

Patch-set: 36
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>