Update patch set 3
Patch Set 3:

Is this patch designed so that UUI server will use the postgres database?
If so, it is not working.
https://logs.onap.org/onap-integration/gating//131444-3/k8s/k8s/kubernetes-status/pod-onap-uui-server-dff7dd44b-hxg72-uui-server-logs.html 
Please do not redeploy in it's current state.
Deploy and test locally and then push amend this patch.

Patch-set: 3
Reviewer: Gerrit User 4405 <4405@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>