Update patch set 2
Patch Set 2:

I tested robot healthcheck tests on the SM environment with and without sidecar.
both are working fine.
If we would enable the sidecar, the benefit would be, 
that we could see the communication between robot and the services in kiali.
So I would abandon this patch.

Patch-set: 2
Reviewer: Gerrit User 2679 <2679@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>