Update patch set 14
Patch Set 14: Code-Review-2

(19 comments)

Hi,
We can review and update this patch but unfortunately this will have to be deferred to the London release. As soon as the London branch is cut for OOM, this will be first in the queue.

@Gareth Roper and Andreas Geissler
Could you guys advise on the changes needed for SM, SecurityContext and Service account?
Thanks

Patch-set: 14
Reviewer: Gerrit User 4405 <4405@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2