Update patch set 15
Patch Set 15:

(11 comments)

Hi Fiachra and and Andreas,

Thanks for the comments. I have resolved some of them.
For the rest ones, they are almost all about using the templates.
Is it possible to merge this one first to include it in K release. Because UUI will release this component in Kohn and users will expect to deploy it with oom.
I have test it locally and everything of UUI works well.

I can submit another patch to do the rest things such as using sevice templates, using a common postgres DB, adding serviceAccount et al.

Thanks

Patch-set: 15
Work-in-progress: false