Update patch set 3
Patch Set 3:

(4 comments)

Just to clarify.  My understanding is that CMPv2 is used by components to get certificates to use on their external interfaces (outside the k8s cluster), not for internal TLS.  Possibly that's made obsolete by the Istio Ingress, but I think it's a separate project to remove it.

Patch-set: 3