Update patch set 3
Patch Set 3: -Code-Review

> Patch Set 3: -Code-Review
> 
> (1 comment)
> 
> I've rebased our codes. Let's see if things are getting OK.
> Btw, I added MSB registration in the "annotations" section as Andreas suggested. In my opinion, MSB has not been announced deprecated officially by ONAP, so it's better for us to mark it as an alternative that's controlled by the switch ".Values.global.msbEnabled".

Thanks Guangrong. I was unsure of the msb requirement in the annotations. If this is needed then so be it.
The last gating run has some issues with smoke tests but may not be related.

Patch-set: 3
Reviewer: Gerrit User 4405 <4405@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0