Update patch set 3
Patch Set 3:

Hi,
the MSB is still an option and the annotation is fine.
But I see in the logs of rule-management pod some SSL related exceptions.
Could you check, whether this can be fixed by configuration or if there is a fix in the pod required:

https://logs.onap.org/onap-integration/gating//133666-3/k8s-teardown/k8s-teardown/kubernetes-status/pod-onap-holmes-rule-mgmt-58d74f448-lcjgj-holmes-rule-mgmt-logs.html

Patch-set: 3
Reviewer: Gerrit User 2679 <2679@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>