Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> I am guessing that the kafka config for ncmp-dmi is ok.
> https://logs.onap.org/onap-integration/gating//133394-9/k8s-teardown/k8s-teardown/kubernetes-status/pod-onap-ncmp-dmi-plugin-c77c46bcc-c46hv-ncmp-dmi-plugin.log
> 
> Can't be sure though.
> Have we though about how the topic names will be passed to the apps them selves?

I have raised this patch just to sync the kafka parameters with other params in cps-core.
So if that is ok then this patch can be merged and meanwhile as per our previous discussion I have made all the kafka topics configurable in our code. Need to do necessary changes in OOM charts as well.But that will be a separate patch. If u agree?

Patch-set: 9