Update patch set 3
Patch Set 3:

Hi Andreas. To be honest the complexity of this chart is beyond me. I had initially made those changes here - https://gerrit.onap.org/r/c/oom/+/133714 but decided to abandon.

This was one of the charts that got temporarily moved out on it's own. I don't think AAI guys or anyone for that matter kept it in good shape/up to date.

The whole chart needs an overhaul but unfortunately I don't have the scope to fully test it as to deploy AAI locally is very time consuming, cassandra issues, dependencies etc.

Patch-set: 3