Update patch set 1
Patch Set 1:

(2 comments)

> Patch Set 1:
> 
> (5 comments)
> 
> Hi,
> generally should all components with an ingress configuration include the istio-ingress account.
> In e.g. the portal-ui I was wondering, if the portal-db is needed, as the DB will nor access the portal. Or both clients and server accounts need to be set?
> 
> Best regards
> Andreas

Hi Andreas, 

for now, yes, I think it makes sense, as you say, to include the istio-ingress serviceaccount for services that contain an ingress configuration - I will update files for this. It may be that some of these may be unnecessary/removable - right now we're trying to identify which components need access to each specific service - as we do not have specific traffic data for everything, for now we are basing on whether they are storing links for these services.

Re portal, I don't know specifically just now. It looks like portal 
- has 4 services - portal-app, portal-cassandra, portal-db, and portal-sdk. 
- has 5 serviceAccounts - portal-app-read, portal-cassandra-read, portal-db-read, portal-sdk-read, portal-widget-read
I would expect that, e.g., the portal-app-read serviceaccount may need to be authorized to read data from/send data to the portal-db service, but unlikely that the portal-db-read serviceaccount will need to read data from the portal-app service.

Patch-set: 1