Update patch set 9
Patch Set 9:

Sorry .. was on vacation last week and missed some comments below.

In London, the intent was to still support the older Biermann version of RESTCONF but to deprecate it.  It is no longer supported in OpenDaylight Chlorine itself, but we took a version of that code in CCSDK and ported it to Chlorine ourselves.  However, I did not install that feature in this submission.  I'll create a new image with that version installed and bump the version and that should clear these issues for now.  

In Montreal, we plan to remove support for the older format to be consistent with the ODL supported version, so at least this oversight helped identify what changes will be needed.

Patch-set: 9