[SDNC] Replace Bierman RESTCONF with RFC 8040
Replaces Bierman draft version of RESTCONF API with
version that is compliant with RFC 8040.
NOTE: This requires corresponding update in SO.  SO team is aware

Issue-ID: SDNC-1847
Change-Id: Id133d6bb83d55d8dac65447e5828b7793069cefb
Signed-off-by: Dan Timoney <dtimoney@att.com>

[SDNC] Replace Bierman RESTCONF with RFC 8040

Replaces Bierman draft version of RESTCONF API with
version that is compliant with RFC 8040.
Updated the required configuration files for SDNC
Update the generic-resource-api configuration file and mount it to
the sdnc pod
NOTE: This requires corresponding update in SO.  SO team is aware

Issue-ID: SDNC-1847
Change-Id: Id133d6bb83d55d8dac65447e5828b7793069cefb
Signed-off-by: Andreas Geissler <andreas-geissler@telekom.de>
Signed-off-by: Dan Timoney <dtimoney@att.com>