[COMMON] Mariadb - Move liveness, readiness and startup probe values to values.yaml
Mariadb fails to come up in resource constrained environments due to the time taken to complete the db initialization.
Most of the time it is more than the startup probe initial delay. Hence the startup probe is failing and mariadb pod keeps on restarting.
This moves the liveness, readiness, and startup probe values from the template file to the `values.yaml` file for better configuration management.

Issue-ID: OOM-3349
Change-Id: I111a426353e3cb3a68c92e4b468c91508e724a8c
Signed-off-by: aravind.est <aravindhan.a@est.tech>